removed unused blocks
