+++ src/diagnostics/syntaxError.ts
@@ -4,7 +4,7 @@ import {isCommandName, isConditionalCommand, isError, isReturn} from '../utils/n
 import {findFirstSibling, getRange, getSiblingNodes} from '../utils/tree-sitter';
 import * as errorCodes from './errorCodes';
 import {createDiagnostic} from './create';
-import {containsRange} from '../workspace-symbol';
+//import {containsRange} from '../workspace-symbol';
 
 // https://github.com/typescript-language-server/typescript-language-server/blob/5a39c1f801ab0cad725a2b8711c0e0d46606a08b/src/diagnostic-queue.ts
 export function getMissingEndSyntaxError(node: SyntaxNode): Diagnostic | null {
@@ -38,60 +38,3 @@ export function getReturnSiblings(node: SyntaxNode) : SyntaxNode[] {
     }
     return results;
 }
-
-export function getUnreachableCodeSyntaxError(node: SyntaxNode): Diagnostic | null {
-    if (!node.isNamed()) return null;
-    //const siblings = findFirstSibling(node, (n) => isConditionalCommand(n) || isReturn(n));
-    //diagnostic.push(...siblings.map(sibling => createDiagnostic(sibling, errorCodes.unreachableCode)))
-    //return null;
-    const returnSibling = findFirstSibling(node, (n: SyntaxNode) => isReturn(n), 'before');
-    if (!returnSibling) return null
-    return findFirstSibling(returnSibling, (n) => !isConditionalCommand(n) && containsRange(getRange(n), getRange(node)), 'after')
-        ? createDiagnostic(node, errorCodes.unreachableCode)
-        : null;
-}
-
-// was -> getMissingEndSyntaxError(node: SyntaxNode): Diagnostic | null
-//if (!isError(node)) return null;
-//return {
-//    severity: DiagnosticSeverity.Error,
-//    code: errorCodes.missingEnd,
-//    message: "Error: Missing end",
-//    range: getRange(node),
-//    source: "fish-lsp",
-//    relatedInformation: getChildNodes(node)
-//        .filter(isBlock)
-//        .map((block) => {
-//            return {
-//                location: { uri: doc.uri, range: getRange(block) },
-//                message: "Potentially missing end",
-//            };
-//        }),
-//};
-
-
-
-
-// const syntaxErrors = nodes.filter(isError).map(n => getChildNodes(n).filter(isBlock)).flat().map(n => n.firstChild || n)
-// const result: Diagnostic[] = []
-// return nodes.filter(isError).map(e => e.firstChild || e)
-//     .map((syntaxError) => {
-//         return Diagnostic.create(
-//             getRange(syntaxError),
-//             `SyntaxError: missing "end" command`,
-//             DiagnosticSeverity.Warning,
-//             errorCodes.missingEnd,
-//             "fish-lsp",
-//             syntaxErrors.map((n) => DiagnosticRelatedInformation.create({uri, range: getRange(n)}, "possibly missing end command"))
-//         )
-//     })
-
-//return Diagnostic.create(
-    //getRange(syntaxError),
-    //`Syntax Error`,
-    //DiagnosticSeverity.Error,
-    //2,
-    //"fish-lsp"
-//);
-
-
